feat(prime_factorization): update functions to use i64 and include types.hpp

+++ header/numbers_theory/prime_factorization.hpp
@@ -3,6 +3,7 @@
 #include <string>
 #include <vector>
 #include <unordered_map>
+#include "types.hpp"
 
-std::vector<int> primeFactors(int n);
-std::unordered_map<int,int> countPrimeFactors(const std::vector<int>& factors);
+std::vector<i64> primeFactors(i64 n);
+std::unordered_map<i64,int> countPrimeFactors(const std::vector<i64>& factors);
